Move ParseRawServiceMessage to tests project
